refactor: depart defaultSongs from recommendMusicPage
