+++ src/components/RecommendMusicPage/index.jsx
@@ -10,60 +10,8 @@ import './style.css';
 import Modal from './Modal';
 import defaultImg from '../../imgs/dummy512.jpg';
 import requestUserFeedback from './requestUserFeedback';
+import defaultSongs from './defaultSongs';
 
-const defaultSongs = [
-  {
-    song_id: 1,
-    song_title: '음악1',
-    artist_name: '가수1',
-    album_title: '앨범1',
-    music_url:
-      'https://p.scdn.co/mp3-preview/6801b88192f9a6dc4ee0c256dfbf72671dc83c7a?cid=0037db4c059245e2a697e420eaf107e0',
-  },
-  {
-    song_id: 2,
-    song_title: '음악2음악2음악2음악2음악2음악2음악2음악2음악2',
-    artist_name:
-      '가수가수가수가수가수가수가수가수가수가수가수가수가수가수가수2',
-    album_title: '앨범2',
-    music_url:
-      'https://p.scdn.co/mp3-preview/bd50f6e6d92ed09aaf38b759ddd96c7164931aa1?cid=0037db4c059245e2a697e420eaf107e0',
-  },
-  {
-    song_id: 3,
-    song_title:
-      '사랑하긴 했었나요 스쳐가는 인연이었나요 짧지 않은 우리 함께했던 시간들이 자꾸 내 마음을 가둬두네',
-    artist_name: '잔미',
-    album_title: '앨범앨범앨범앨범앨범앨범앨범앨범3',
-    music_url:
-      'https://p.scdn.co/mp3-preview/5274582d3f3f34e82fb83e814dcad59182c92ae4?cid=0037db4c059245e2a697e420eaf107e0',
-  },
-  {
-    song_id: 4,
-    song_title: '음악4',
-    artist_name:
-      '사랑하긴 했었나요 스쳐가는 인연이었나요 짧지 않은 우리 함께했던 시간들이 자꾸 내 마음을 가둬두네',
-    album_title: '앨범4',
-    music_url:
-      'https://p.scdn.co/mp3-preview/a54a328016674ea6da346af84c6c06c6536195c8?cid=0037db4c059245e2a697e420eaf107e0',
-  },
-  {
-    song_id: 5,
-    song_title: '음악5',
-    artist_name: '가수5',
-    album_title: '앨범5',
-    music_url:
-      'https://p.scdn.co/mp3-preview/1aaf30001acf34f7c6c0d9aee23c354d05923fc4?cid=0037db4c059245e2a697e420eaf107e0',
-  },
-  {
-    song_id: 6,
-    song_title: '음악6',
-    artist_name: '가수6',
-    album_title: '앨범6',
-    music_url:
-      'https://p.scdn.co/mp3-preview/1eff5d6411ae2c65990552133729f3e907c4a793?cid=0037db4c059245e2a697e420eaf107e0',
-  },
-];
 const defaultId = 'NULL';
 
 function RecommendMusicPage() {
